feat: Add structured handlers for the document operations.
For the operations happenining while collaborating (editing) the text document, it will be sent over websocket messages between the clients and the server. It will be a json string where there is an attribute called  that will specify the type of the operation. Each operation type will have its associated attributes which will be extracted in each operation handler.